+++ pubmodules/pubModulesManager.js
@@ -35,6 +35,9 @@ class PubModulesManager {
 
         this.sms = undefined;
         this.smsRoute = undefined;
+        
+        this.voice = undefined;
+        this.voiceRoute = undefined;
 
         this.mqttTest = undefined;
         this.mqttTestRoute = undefined;
@@ -97,6 +100,10 @@ class PubModulesManager {
         if (this.smsRoute) {
             app.use('/modules/sms', this.smsRoute);
             winston.info("PubModulesManager smsRoute controller loaded");
+        }
+        if (this.voiceRoute) {
+            app.use('/modules/voice', this.voiceRoute);
+            winston.info("PubModulesManager voiceRoute controller loaded");
         }
         if (this.mqttTestRoute) {
             app.use('/modules/mqttTest', this.mqttTestRoute);
@@ -335,6 +342,25 @@ class PubModulesManager {
                 winston.info("PubModulesManager init apps module not found ");
             } else {
                 winston.info("PubModulesManager error initializing init apps module", err);
+            }
+        }
+
+        if (process.env.VOICE_TOKEN === process.env.VOICE_SECRET) {
+            try {
+                this.voice = require('./voice');
+                winston.info("this.voice: " + this.voice);
+                this.voice.listener.listen(config);
+    
+                this.voiceRoute = this.voice.voiceRoute;
+    
+                winston.info("PubModulesManager initialized apps (voice).")
+            } catch(err) {
+                console.log("\n Unable to start voice connector: ", err);
+                if (err.code == 'MODULE_NOT_FOUND') {
+                    winston.info("PubModulesManager init apps module not found ");
+                } else {
+                    winston.info("PubModulesManager error initializing init apps module", err);
+                }
             }
         }