+++ routes/kb.js
@@ -13,6 +13,7 @@ const Sitemapper = require('sitemapper');
 var mongoose = require('mongoose');
 const faq = require('../models/faq');
 const faq_kb = require('../models/faq_kb');
+let Integration = require('../models/integrations');
 
 const { MODELS_MULTIPLIER } = require('../utils/aiUtils');
 
@@ -32,8 +33,8 @@ jobManager.connectAndStartPublisher(() => {
 })
 
 let default_preview_settings = {
-  model: 'gpt-3.5-turbo',
-  max_tokens: 128,
+  model: 'gpt-4o',
+  max_tokens: 256,
   temperature: 0.7,
   top_k: 4,
   //context: "You are an awesome AI Assistant."
@@ -53,6 +54,10 @@ router.post('/scrape/single', async (req, res) => {
 
   let data = req.body;
   winston.debug("/scrape/single data: ", data);
+
+  if (!data.scrape_type) {
+    data.scrape_type = 1;
+  }
 
   let namespaces = await Namespace.find({ id_project: project_id }).catch((err) => {
     winston.error("find namespaces error: ", err)
@@ -92,6 +97,11 @@ router.post('/scrape/single', async (req, res) => {
 
       if (kb.content) {
         json.content = kb.content;
+      }
+
+      json.scrape_type = 1;
+      if (data.scrape_type) {
+        json.scrape_type = data.scrape_type;
       }
 
       startScrape(json).then((response) => {
@@ -177,20 +187,22 @@ router.post('/qa', async (req, res) => {
   }
 
   let namespaceIds = namespaces.map(namespace => namespace.id);
-
   if (!namespaceIds.includes(data.namespace)) {
     return res.status(403).send({ success: false, error: "Not allowed. The namespace does not belong to the current project." })
   }
-
+  
   winston.debug("/qa data: ", data);
 
   if (!data.gptkey) {
-    let gptkey = process.env.GPTKEY;
+    let gptkey = await getKeyFromIntegrations(project_id);
+    if (!gptkey) {
+      gptkey = process.env.GPTKEY;
+      publicKey = true;
+    }
     if (!gptkey) {
       return res.status(403).send({ success: false, error: "GPT apikey undefined" })
     }
     data.gptkey = gptkey;
-    publicKey = true;
   }
 
   let obj = { createdAt: new Date() };
@@ -204,9 +216,17 @@ router.post('/qa', async (req, res) => {
   }
 
   // Check if "Advanced Mode" is active. In such case the default_context must be not appended
-  if (data.system_context) {
-    data.system_context = data.system_context + " \n" + default_context;
+  if (!data.advanced_context) {
+    if (data.system_context) {
+      data.system_context = data.system_context + " \n" + default_context;
+    } else {
+      data.system_context = default_context;
+    }
   }
+
+  // if (process.env.NODE_ENV === 'test') {
+  //   return res.status(200).send({ success: true, message: "Question skipped in test environment"});
+  // }
 
   openaiService.askNamespace(data).then((resp) => {
     winston.debug("qa resp: ", resp.data);
@@ -220,16 +240,18 @@ router.post('/qa', async (req, res) => {
 
     KB.findById(id, (err, resource) => {
 
-      let multiplier = MODELS_MULTIPLIER[data.model];
-      if (!multiplier) {
-        multiplier = 1;
-        winston.info("No multiplier found for AI model")
+      if (publicKey === true) {
+        let multiplier = MODELS_MULTIPLIER[data.model];
+        if (!multiplier) {
+          multiplier = 1;
+          winston.info("No multiplier found for AI model")
+        }
+        obj.multiplier = multiplier;
+        obj.tokens = answer.prompt_token_size;
+  
+        let incremented_key = quoteManager.incrementTokenCount(req.project, obj);
+        winston.verbose("incremented_key: ", incremented_key);
       }
-      obj.multiplier = multiplier;
-      obj.tokens = answer.prompt_token_size;
-
-      let incremented_key = quoteManager.incrementTokenCount(req.project, obj);
-      winston.verbose("incremented_key: ", incremented_key);
 
       if (err) {
         winston.error("Unable to find resource with id " + id + " in namespace " + answer.namespace + ". The standard answer is returned.")
@@ -504,7 +526,7 @@ router.post('/namespace', async (req, res) => {
   let quoteManager = req.app.get('quote_manager');
   let limits = await quoteManager.getPlanLimits(req.project);
   let ns_limit = limits.namespace;
-  console.log("Limit of namespaces for current plan " + ns_limit);
+  //console.log("Limit of namespaces for current plan " + ns_limit);
 
   if (namespaces.length >= ns_limit) {
     return res.status(403).send({ success: false, error: "Maximum number of resources reached for the current plan", plan_limit: ns_limit });
@@ -1184,6 +1206,22 @@ async function startScrape(data) {
       winston.error("singleScrape err: ", err);
       reject(err);
     })
+  })
+}
+
+async function getKeyFromIntegrations(project_id) {
+
+  return new Promise( async (resolve) => {
+
+    let integration = await Integration.findOne({ id_project: project_id, name: 'openai' }).catch((err) => {
+      winston.error("Unable to find openai integration for the current project " + project_id);
+      resolve(null);
+    })
+    if (integration && integration.value && integration.value.apikey) {
+      resolve(integration.value.apikey);
+    } else {
+      resolve(null);
+    }
   })
 }
 /**